requirements.txt
behave==1.2.6
certifi==2024.8.30
charset-normalizer==3.4.0
coverage==7.6.3
execnet==2.1.1
Faker==30.6.0
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==3.0.2
packaging==24.1
parse==1.20.2
parse_type==0.6.4
pluggy==1.5.0
pytest==8.3.3
pytest-bdd==7.3.0
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
six==1.16.0
typing_extensions==4.12.2
urllib3==2.2.3
